fix: response to review comments
This set of changes includes the following:

1. Renaming short variable names to be longer and more understandable.
2. Not using the Status.ScheduleStartTime for the pod start time, but instead
adding a new field. This previous field was there for a different purpose.
3. Creating named identifiers for resource types that can be shared in the
jgf module along with others that use the same relations / vertex types.
4. Removing comments that are not necessary.
5. Changing JGF types from int to int64 that warrant it.
6. Spelling mistakes, etc.
7. Removing need to write jobspec to temporary file (we just need string)

The JGF and utils modules need some additional looking - specifically I am worried
that the paths->containment is not set, and sometimes the name reflects the index
of the overall graph (global) and other times the index of the resource type. I
think we likely want the latter for the inner name, but I am not sure in practice
that fluxion is using it (internally). I am pushing these changes to assess testing,
etc., and will update the PR as needed. There could also have been changes to
upstream since the PR was opened that warrant additional fixes.

+++ src/fluence/jgf/types.go
@@ -38,12 +38,12 @@ type nodeMetadata struct {
 	Type       string            `json:"type"`
 	Basename   string            `json:"basename"`
 	Name       string            `json:"name"`
-	Id         int               `json:"id"`
-	Uniq_id    int               `json:"uniq_id"`
-	Rank       int               `json:"rank,omitempty"`
+	Id         int64             `json:"id"`
+	Uniq_id    int64             `json:"uniq_id"`
+	Rank       int64             `json:"rank,omitempty"`
 	Exclusive  bool              `json:"exclusive"`
 	Unit       string            `json:"unit"`
-	Size       int               `json:"size"`
+	Size       int64             `json:"size"`
 	Paths      map[string]string `json:"paths,omitempty"`
 	Properties map[string]string `json:"properties,omitempty"`
 }
@@ -57,6 +57,6 @@ type graph struct {
 
 type Fluxjgf struct {
 	Graph    graph           `json:"graph"`
-	Elements int             `json:"-"`
+	Elements int64           `json:"-"`
 	NodeMap  map[string]node `json:"-"`
 }
